app-web/requirements.txt
python-docx==0.8.11
flask==2.0.2
flask_navigation==0.2.0
pyyaml==5.4
redis==4.0.2
exchangelib==4.6.0
python-ldap==3.4.0